+++ src/DivideContract/distriute.sol
@@ -1,25 +1,24 @@
 pragma solidity ^0.8.0;
+
 import "../like/MyToken.sol";
 
 contract WalletDistributor {
     address public owner;
     IERC20 public immutable myToken;
 
-   constructor(address _owner, address _myToken) {
+    constructor(address _owner, address _myToken) {
         owner = _owner;
         myToken = IERC20(_myToken);
     }
 
-    function distribute(address[] memory recipients,uint amount) public payable {
+    function distribute(address[] memory recipients, uint256 amount) public payable {
         require(msg.sender == owner, "Only owner can distribute funds");
         require(recipients.length > 0, "No recipients specified");
         uint256 amountPerRecipient = amount / recipients.length;
         require(amountPerRecipient > 0, "Amount per recipient is too low");
 
-        for (uint i = 0; i < recipients.length; i++) {
-           myToken.transferFrom(owner,recipients[i],amountPerRecipient);
+        for (uint256 i = 0; i < recipients.length; i++) {
+            myToken.transferFrom(owner, recipients[i], amountPerRecipient);
         }
     }
-
-   
 }
